+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `dvd-dashboard-ctl.sh` script for start/stop/restart/status of web dashboard
+
 ## [1.1.0] - 2025-01-05
 
 ### Added